migrated communities screen
